Expose BaseDocument and TypeDocument
Exposing these to allow a future database API change for replaces that
requires being able to pass in an etag, which is on the BaseDocument.

This PR is just renames and moves.  There will be followups, but keeping
this one to no-op changes.
