+++ internal/database/document.go
@@ -14,391 +14,8 @@
 
 package database
 
-import (
-	"time"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	azcorearm "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos"
-	"github.com/google/uuid"
-
-	"github.com/Azure/ARO-HCP/internal/api"
-	"github.com/Azure/ARO-HCP/internal/api/arm"
-	"github.com/Azure/ARO-HCP/internal/ocm"
-)
-
-// baseDocument includes fields common to all container items.
-type baseDocument struct {
-	ID string `json:"id,omitempty"`
-
-	// https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/time-to-live
-	TimeToLive int `json:"ttl,omitempty"`
-
-	// System-defined properties generated by Cosmos DB
-	CosmosResourceID  string      `json:"_rid,omitempty"`
-	CosmosSelf        string      `json:"_self,omitempty"`
-	CosmosETag        azcore.ETag `json:"_etag,omitempty"`
-	CosmosAttachments string      `json:"_attachments,omitempty"`
-	CosmosTimestamp   int         `json:"_ts,omitempty"`
-}
-
-// newBaseDocument returns a baseDocument with a unique ID.
-func newBaseDocument() baseDocument {
-	return baseDocument{ID: uuid.New().String()}
-}
-
 // DocumentProperties is an interface for types that can serve as
-// typedDocument.Properties.
+// TypedDocument.Properties.
 type DocumentProperties interface {
 	GetValidTypes() []string
-}
-
-// BillingDocument records timestamps of Hosted Control Plane OpenShift cluster
-// creation and deletion for the purpose of customer billing.
-type BillingDocument struct {
-	baseDocument
-
-	// The cluster creation time represents the time when the cluster was provisioned successfully
-	CreationTime time.Time `json:"creationTime,omitempty"`
-	// The cluster deletion time
-	DeletionTime *time.Time `json:"deletionTime,omitempty"`
-
-	// The location of the HCP cluster
-	Location string `json:"location,omitempty"`
-	// The tenant ID of the HCP cluster
-	TenantID string `json:"tenantId,omitempty"`
-	// The subscription ID of the HCP cluster (also the partition key)
-	SubscriptionID string `json:"subscriptionId,omitempty"`
-	// The HCP cluster ARM resource ID
-	ResourceID *azcorearm.ResourceID `json:"resourceId,omitempty"`
-	// The ARM resource ID of the managed resource group of the HCP cluster
-	ManagedResourceGroup string `json:"managedResourceGroup,omitempty"`
-}
-
-func NewBillingDocument(resourceID *azcorearm.ResourceID) *BillingDocument {
-	return &BillingDocument{
-		baseDocument:   newBaseDocument(),
-		SubscriptionID: resourceID.SubscriptionID,
-		ResourceID:     resourceID,
-	}
-}
-
-// BillingDocumentPatchOperations represents a patch request for a BillingDocument.
-type BillingDocumentPatchOperations struct {
-	azcosmos.PatchOperations
-}
-
-// SetDeletionTime appends a set operation for the DeletionTime field.
-func (p *BillingDocumentPatchOperations) SetDeletionTime(deletionTime time.Time) {
-	p.AppendSet("/deletionTime", deletionTime)
-}
-
-// ResourceDocument captures the mapping of an Azure resource ID
-// to an internal resource ID (the OCM API path), as well as any
-// ARM-specific metadata for the resource.
-type ResourceDocument struct {
-	ResourceID        *azcorearm.ResourceID       `json:"resourceId,omitempty"`
-	InternalID        ocm.InternalID              `json:"internalId,omitempty"`
-	ActiveOperationID string                      `json:"activeOperationId,omitempty"`
-	ProvisioningState arm.ProvisioningState       `json:"provisioningState,omitempty"`
-	Identity          *arm.ManagedServiceIdentity `json:"identity,omitempty"`
-	SystemData        *arm.SystemData             `json:"systemData,omitempty"`
-	Tags              map[string]string           `json:"tags,omitempty"`
-
-	CustomerDesiredState map[string]any `json:"customerDesiredState"`
-	ServiceProviderState map[string]any `json:"serviceProviderState"`
-}
-
-func NewResourceDocument(resourceID *azcorearm.ResourceID) *ResourceDocument {
-	return &ResourceDocument{
-		ResourceID: resourceID,
-	}
-}
-
-// GetValidTypes returns the valid resource types for a ResourceDocument.
-func (doc ResourceDocument) GetValidTypes() []string {
-	return []string{
-		api.ClusterResourceType.String(),
-		api.NodePoolResourceType.String(),
-		api.ExternalAuthResourceType.String(),
-	}
-}
-
-// ResourceDocumentStateFilter is used to remove unknown fields from ResourceDocumentProperties.
-// Long-term, we want to reach a point where we store different types so we have full type-safety
-// throughout the stack.
-// Short-term, we want a low-touch modification that makes it safe to store new fields.
-type ResourceDocumentStateFilter interface {
-	// RemoveUnknownFields checks the customerDesiredState and serviceProviderState and removes unknown fields.
-	// The simplest implementation is "remove everything" and the next simplest is round-tripping through JSON.
-	RemoveUnknownFields(*ResourceDocument) error
-}
-
-const (
-	ResourceDocumentJSONPathResourceID        = typedDocumentJSONPathProperties + "/resourceId"
-	ResourceDocumentJSONPathInternalID        = typedDocumentJSONPathProperties + "/internalId"
-	ResourceDocumentJSONPathActiveOperationID = typedDocumentJSONPathProperties + "/activeOperationId"
-	ResourceDocumentJSONPathProvisioningState = typedDocumentJSONPathProperties + "/provisioningState"
-	ResourceDocumentJSONPathIdentity          = typedDocumentJSONPathProperties + "/identity"
-	ResourceDocumentJSONPathSystemData        = typedDocumentJSONPathProperties + "/systemData"
-	ResourceDocumentJSONPathTags              = typedDocumentJSONPathProperties + "/tags"
-)
-
-// ResourceDocumentPatchOperations represents a patch request for a ResourceDocument.
-type ResourceDocumentPatchOperations struct {
-	azcosmos.PatchOperations
-}
-
-// SetActiveOperationID appends a set or remove operation for the ActiveOperationID
-// field, depending on whether activeOperationID is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *ResourceDocumentPatchOperations) SetActiveOperationID(activeOperationID *string) {
-	if activeOperationID != nil {
-		p.AppendSet(ResourceDocumentJSONPathActiveOperationID, *activeOperationID)
-	} else {
-		p.AppendRemove(ResourceDocumentJSONPathActiveOperationID)
-	}
-}
-
-// SetProvisioningState appends a set operation for the ProvisioningState field.
-func (p *ResourceDocumentPatchOperations) SetProvisioningState(provisioningState arm.ProvisioningState) {
-	p.AppendSet(ResourceDocumentJSONPathProvisioningState, provisioningState)
-}
-
-// SetIdentity appends a set or remove operation for the Identity field,
-// depending on whether identity is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *ResourceDocumentPatchOperations) SetIdentity(identity *arm.ManagedServiceIdentity) {
-	if identity != nil {
-		p.AppendSet(ResourceDocumentJSONPathIdentity, identity)
-	} else {
-		p.AppendRemove(ResourceDocumentJSONPathIdentity)
-	}
-}
-
-// SetSystemData appends a set or remove operation for the SystemData field,
-// depending on whether systemData is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *ResourceDocumentPatchOperations) SetSystemData(systemData *arm.SystemData) {
-	if systemData != nil {
-		p.AppendSet(ResourceDocumentJSONPathSystemData, systemData)
-	} else {
-		p.AppendRemove(ResourceDocumentJSONPathSystemData)
-	}
-}
-
-// SetTags appends a set or remove operation for the Tags field, depending on
-// whether tags is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *ResourceDocumentPatchOperations) SetTags(tags map[string]string) {
-	if tags != nil {
-		p.AppendSet(ResourceDocumentJSONPathTags, tags)
-	} else {
-		p.AppendRemove(ResourceDocumentJSONPathTags)
-	}
-}
-
-type OperationRequest string
-
-const (
-	OperationRequestCreate OperationRequest = "Create"
-	OperationRequestUpdate OperationRequest = "Update"
-	OperationRequestDelete OperationRequest = "Delete"
-
-	// These are for POST actions on resources.
-	OperationRequestRequestCredential OperationRequest = "RequestCredential"
-	OperationRequestRevokeCredentials OperationRequest = "RevokeCredentials"
-)
-
-// OperationResourceType is an artificial resource type for OperationDocuments
-// in Cosmos DB. It omits the location segment from actual operation endpoints.
-var OperationResourceType = azcorearm.NewResourceType(api.ProviderNamespace, api.OperationStatusResourceTypeName)
-
-// OperationDocument tracks an asynchronous operation.
-type OperationDocument struct {
-	// TenantID is the tenant ID of the client that requested the operation
-	TenantID string `json:"tenantId,omitempty"`
-	// ClientID is the object ID of the client that requested the operation
-	ClientID string `json:"clientId,omitempty"`
-	// Request is the type of asynchronous operation requested
-	Request OperationRequest `json:"request,omitempty"`
-	// ExternalID is the Azure resource ID of the cluster or node pool
-	ExternalID *azcorearm.ResourceID `json:"externalId,omitempty"`
-	// InternalID is the Cluster Service resource identifier in the form of a URL path
-	InternalID ocm.InternalID `json:"internalId,omitempty"`
-	// OperationID is the Azure resource ID of the operation status (may be nil if the
-	// operation was implicit, such as deleting a child resource along with the parent)
-	OperationID *azcorearm.ResourceID `json:"operationId,omitempty"`
-	// ClientRequestID is provided by the "x-ms-client-request-id" request header
-	ClientRequestID string `json:"clientRequestId,omitempty"`
-	// CorrelationRequstID is provided by the "x-ms-correlation-request-id" request header
-	CorrelationRequestID string `json:"correlationRequestId,omitempty"`
-	// NotificationURI is provided by the Azure-AsyncNotificationUri header if the
-	// Async Operation Callbacks ARM feature is enabled
-	NotificationURI string `json:"notificationUri,omitempty"`
-
-	// StartTime marks the start of the operation
-	StartTime time.Time `json:"startTime,omitempty"`
-	// LastTransitionTime marks the most recent state change
-	LastTransitionTime time.Time `json:"lastTransitionTime,omitempty"`
-	// Status is the current operation status, using the same set of values
-	// as the resource's provisioning state
-	Status arm.ProvisioningState `json:"status,omitempty"`
-	// Error is an OData error, present when Status is "Failed" or "Canceled"
-	Error *arm.CloudErrorBody `json:"error,omitempty"`
-}
-
-func NewOperationDocument(request OperationRequest, externalID *azcorearm.ResourceID, internalID ocm.InternalID, correlationData *arm.CorrelationData) *OperationDocument {
-	now := time.Now().UTC()
-
-	doc := &OperationDocument{
-		Request:            request,
-		ExternalID:         externalID,
-		InternalID:         internalID,
-		StartTime:          now,
-		LastTransitionTime: now,
-		Status:             arm.ProvisioningStateAccepted,
-	}
-
-	if correlationData != nil {
-		doc.ClientRequestID = correlationData.ClientRequestID
-		doc.CorrelationRequestID = correlationData.CorrelationRequestID
-	}
-
-	// When deleting, set Status directly to ProvisioningStateDeleting
-	// so any further deletion requests are rejected with 409 Conflict.
-	if request == OperationRequestDelete {
-		doc.Status = arm.ProvisioningStateDeleting
-	}
-
-	return doc
-}
-
-// GetValidTypes returns the valid resource types for an OperationDocument.
-func (doc OperationDocument) GetValidTypes() []string {
-	return []string{OperationResourceType.String()}
-}
-
-// ToStatus converts an OperationDocument to the ARM operation status format.
-func (doc *OperationDocument) ToStatus() *arm.Operation {
-	operation := &arm.Operation{
-		ID:        doc.OperationID,
-		Name:      doc.OperationID.Name,
-		Status:    doc.Status,
-		StartTime: &doc.StartTime,
-		Error:     doc.Error,
-	}
-
-	if doc.Status.IsTerminal() {
-		operation.EndTime = &doc.LastTransitionTime
-	}
-
-	return operation
-}
-
-// UpdateStatus conditionally updates the document if the status given differs
-// from the status already present. If so, it sets the Status and Error fields
-// to the values given, updates the LastTransitionTime, and returns true. This
-// is intended to be used with DBClient.UpdateOperationDoc.
-func (doc *OperationDocument) UpdateStatus(status arm.ProvisioningState, err *arm.CloudErrorBody) bool {
-	if doc.Status != status {
-		doc.LastTransitionTime = time.Now().UTC()
-		doc.Status = status
-		doc.Error = err
-		return true
-	}
-	return false
-}
-
-// OperationDocumentPatchOperations represents a patch request for an OperationDocument.
-type OperationDocumentPatchOperations struct {
-	azcosmos.PatchOperations
-}
-
-const (
-	OperationDocumentJSONPathTenantID           = typedDocumentJSONPathProperties + "/tenantId"
-	OperationDocumentJSONPathClientID           = typedDocumentJSONPathProperties + "/clientId"
-	OperationDocumentJSONPathRequest            = typedDocumentJSONPathProperties + "/request"
-	OperationDocumentJSONPathExternalID         = typedDocumentJSONPathProperties + "/externalId"
-	OperationDocumentJSONPathInternalID         = typedDocumentJSONPathProperties + "/internalId"
-	OperationDocumentJSONPathOperationID        = typedDocumentJSONPathProperties + "/operationId"
-	OperationDocumentJSONPathNotificationURI    = typedDocumentJSONPathProperties + "/notificationUri"
-	OperationDocumentJSONPathStartTime          = typedDocumentJSONPathProperties + "/startTime"
-	OperationDocumentJSONPathLastTransitionTime = typedDocumentJSONPathProperties + "/lastTransitionTime"
-	OperationDocumentJSONPathStatus             = typedDocumentJSONPathProperties + "/status"
-	OperationDocumentJSONPathError              = typedDocumentJSONPathProperties + "/error"
-)
-
-// SetTenantID appends a set operation for the TenantID field.
-func (p *OperationDocumentPatchOperations) SetTenantID(tenantID string) {
-	p.AppendSet(OperationDocumentJSONPathTenantID, tenantID)
-}
-
-// SetClientID appends a set operation for the ClientID field.
-func (p *OperationDocumentPatchOperations) SetClientID(clientID string) {
-	p.AppendSet(OperationDocumentJSONPathClientID, clientID)
-}
-
-// SetOperationID appends a set or remove operation for the OperationID field,
-// depending on whether operationID is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *OperationDocumentPatchOperations) SetOperationID(operationID *azcorearm.ResourceID) {
-	if operationID != nil {
-		p.AppendSet(OperationDocumentJSONPathOperationID, operationID)
-	} else {
-		p.AppendRemove(OperationDocumentJSONPathOperationID)
-	}
-}
-
-// SetNotificationURI appends a set or remove operation for the NotificationURI field,
-// depending on whether notificationURI is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *OperationDocumentPatchOperations) SetNotificationURI(notificationURI *string) {
-	if notificationURI != nil {
-		p.AppendSet(OperationDocumentJSONPathNotificationURI, *notificationURI)
-	} else {
-		p.AppendRemove(OperationDocumentJSONPathNotificationURI)
-	}
-}
-
-// SetLastTransitionTime appends a set operation for the LastTransitionTime field.
-func (p *OperationDocumentPatchOperations) SetLastTransitionTime(lastTransitionTime time.Time) {
-	p.AppendSet(OperationDocumentJSONPathLastTransitionTime, lastTransitionTime)
-}
-
-// SetStatus appends a set operation for the Status field.
-func (p *OperationDocumentPatchOperations) SetStatus(status arm.ProvisioningState) {
-	p.AppendSet(OperationDocumentJSONPathStatus, status)
-}
-
-// SetError appends a set or remove operation for the Error field,
-// depending on whether err is nil.
-//
-// Be careful when appending a remove patch operation as it is NOT idempotent.
-// If the field to remove is not present in the Cosmos DB document, the entire
-// patch request will fail with a "400 Bad Request" status code.
-func (p *OperationDocumentPatchOperations) SetError(err *arm.CloudErrorBody) {
-	if err != nil {
-		p.AppendSet(OperationDocumentJSONPathError, err)
-	} else {
-		p.AppendRemove(OperationDocumentJSONPathError)
-	}
 }
