+++ internal/database/lock.go
@@ -55,7 +55,7 @@ type LockClient struct {
 // lockDocument implements a global distributed lock.
 // Its contents should be opaque outside of LockClient.
 type lockDocument struct {
-	baseDocument
+	BaseDocument
 	Owner string `json:"owner,omitempty"`
 	TTL   int32  `json:"ttl,omitempty"`
 }
@@ -141,7 +141,7 @@ func (c *LockClient) AcquireLock(ctx context.Context, id string, timeout *time.D
 // is already taken, it returns a nil azcosmos.ItemResponse and no error.
 func (c *LockClient) TryAcquireLock(ctx context.Context, id string) (*azcosmos.ItemResponse, error) {
 	doc := &lockDocument{
-		baseDocument: baseDocument{ID: id},
+		BaseDocument: BaseDocument{ID: id},
 		Owner:        c.name,
 		TTL:          c.defaultTimeToLive,
 	}
