+++ internal/database/types_basedocument.go
@@ -1,0 +1,40 @@
+// Copyright 2025 Microsoft Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package database
+
+import (
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/google/uuid"
+)
+
+// BaseDocument includes fields common to all container items.
+type BaseDocument struct {
+	ID string `json:"id,omitempty"`
+
+	// https://learn.microsoft.com/en-us/azure/cosmos-db/nosql/time-to-live
+	TimeToLive int `json:"ttl,omitempty"`
+
+	// System-defined properties generated by Cosmos DB
+	CosmosResourceID  string      `json:"_rid,omitempty"`
+	CosmosSelf        string      `json:"_self,omitempty"`
+	CosmosETag        azcore.ETag `json:"_etag,omitempty"`
+	CosmosAttachments string      `json:"_attachments,omitempty"`
+	CosmosTimestamp   int         `json:"_ts,omitempty"`
+}
+
+// newBaseDocument returns a BaseDocument with a unique ID.
+func newBaseDocument() BaseDocument {
+	return BaseDocument{ID: uuid.New().String()}
+}
